+++ mod-source-record-manager-server/src/main/java/org/folio/rest/impl/ModTenantAPI.java
@@ -4,15 +4,20 @@ import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
+import io.vertx.core.Vertx;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.sql.UpdateResult;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.folio.dataimport.util.OkapiConnectionParams;
 import org.folio.rest.annotations.Validate;
 import org.folio.rest.jaxrs.model.TenantAttributes;
 import org.folio.rest.persist.PostgresClient;
 import org.folio.rest.tools.utils.TenantTool;
+import org.folio.services.MappingRuleService;
+import org.folio.spring.SpringContextUtil;
+import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.ws.rs.core.Response;
 import java.io.IOException;
@@ -35,30 +40,42 @@ public class ModTenantAPI extends TenantAPI {
   public static final String MODULE_PLACEHOLDER = "${mymodule}";
 
 
+  @Autowired
+  private MappingRuleService mappingRuleService;
+
+  public ModTenantAPI() {
+    SpringContextUtil.autowireDependencies(this, Vertx.currentContext());
+  }
+
   @Validate
   @Override
-  public void postTenant(TenantAttributes entity, Map<String, String> headers, Handler<AsyncResult<Response>> handlers, Context context) {
-    super.postTenant(entity, headers, ar -> {
-      if (ar.failed()) {
-        handlers.handle(ar);
+  public void postTenant(TenantAttributes entity, Map<String, String> headers, Handler<AsyncResult<Response>> handler, Context context) {
+    super.postTenant(entity, headers, postTenantAr -> {
+      if (postTenantAr.failed()) {
+        handler.handle(postTenantAr);
       } else {
-        setSequencesPermissionForDbUser(headers, context)
-          .compose(permissionAr ->  setupTestData(headers, context))
-          .setHandler(event -> handlers.handle(ar));
+        OkapiConnectionParams okapiConnectionParams = new OkapiConnectionParams(headers, context.owner());
+        String tenantId = TenantTool.calculateTenantId(okapiConnectionParams.getTenantId());
+        setSequencesPermissionForDbUser(context, tenantId)
+          .compose(permissionAr -> setupTestData(context, tenantId))
+          .compose(ar -> mappingRuleService.saveDefaultRules(tenantId))
+          .setHandler(event -> handler.handle(postTenantAr));
       }
     }, context);
   }
 
-  private Future<UpdateResult> setSequencesPermissionForDbUser(Map<String, String> headers, Context context) {
+  private Future<UpdateResult> setSequencesPermissionForDbUser(Context context, String tenantId) {
     Future<UpdateResult> future = Future.future();
-    String dbSchemaName = new StringBuilder(TenantTool.calculateTenantId(headers.get("x-okapi-tenant")))
-      .append('_').append(PostgresClient.getModuleName()).toString();
+    String dbSchemaName = new StringBuilder()
+      .append(tenantId)
+      .append('_')
+      .append(PostgresClient.getModuleName()).toString();
     String preparedSql = String.format(GRANT_SEQUENCES_PERMISSION_PATTERN, dbSchemaName, dbSchemaName);
     PostgresClient.getInstance(context.owner()).execute(preparedSql, future);
     return future;
   }
 
-  private Future<List<String>> setupTestData(Map<String, String> headers, Context context) {
+  private Future<List<String>> setupTestData(Context context, String tenantId) {
     try {
       if (!Boolean.TRUE.equals(Boolean.valueOf(System.getenv(TEST_MODE)))) {
         LOGGER.info("Test data was not initialized.");
@@ -73,7 +90,6 @@ public class ModTenantAPI extends TenantAPI {
       if (StringUtils.isBlank(sqlScript)) {
         return Future.succeededFuture();
       }
-      String tenantId = TenantTool.calculateTenantId(headers.get("x-okapi-tenant"));
       String moduleName = PostgresClient.getModuleName();
 
       sqlScript = sqlScript.replace(TENANT_PLACEHOLDER, tenantId).replace(MODULE_PLACEHOLDER, moduleName);