+++ NEWS.md
@@ -1,6 +1,7 @@
 ## 2019-09-30 v1.7.0-SNAPSHOT
 * Wrote documentation for Marc-to-Instance mapping processor
 * Added order of the record in importing file
+* Create CLI way for individual tenant to update the default MARC-to-Instance map
 
 ## 2019-09-09 v1.6.1
  * Added instance-type settings loading for mapping